updated the paymentOption service
